- rpm.org: size_t not int.

+++ lib/rpmps.c
@@ -239,7 +239,7 @@ const char * rpmProblemString(const rpmProblem prob)
     const char * altNEVR = (prob->altNEVR ? prob->altNEVR : "? ?altNEVR?");
 /*@observer@*/
     const char * str1 = (prob->str1 ? prob->str1 : N_("different"));
-    int nb =	strlen(pkgNEVR) + strlen(str1) + strlen(altNEVR) + 100;
+    size_t nb =	strlen(pkgNEVR) + strlen(str1) + strlen(altNEVR) + 100;
     char * buf = xmalloc(nb+1);
     int rc;
